requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
python-dotenv==1.0.0
python-docx==1.1.2
PyMuPDF==1.24.7
spacy==3.7.5

https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl